requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==6.7
cycler==0.10.0
DateTime==4.3
flake8==3.8.3
Flask==1.1.2
idna==2.7
itsdangerous==0.24
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==0.23
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
Pillow==7.2.0
plaid-python==5.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.20.0
selenium==3.141.0
six==1.15.0
urllib3==1.24.3
Werkzeug==1.0.1
wordcloud==1.8.0
zope.interface==5.1.0
